Moved MessageLogging non-API headers to src directory
Headers not needed outside of the package moved for src or to
the only package that uses them.
