+++ FWCore/MessageLogger/src/SilentMLscribe.h
