+++ FWCore/MessageService/src/ConfigurationHandshake.h
