+++ FWCore/MessageService/src/ELadministrator.h
@@ -52,7 +52,7 @@
 #include "FWCore/MessageLogger/interface/ELlist.h"
 #include "FWCore/MessageLogger/interface/ELseverityLevel.h"
 #include "FWCore/MessageLogger/interface/ErrorObj.h"
-#include "FWCore/MessageService/interface/ELdestination.h"
+#include "FWCore/MessageService/src/ELdestination.h"
 #include "FWCore/Utilities/interface/propagate_const.h"
 
 #include <memory>
