+++ FWCore/MessageService/src/MessageLoggerDefaults.cc
@@ -16,7 +16,7 @@
 //
 // ----------------------------------------------------------------------
 
-#include "FWCore/MessageService/interface/MessageLoggerDefaults.h"
+#include "FWCore/MessageService/src/MessageLoggerDefaults.h"
 
 namespace edm {
   namespace service {
